Implementing offer methods

+++ src/main/java/com/atrastones/shop/model/entity/Offer.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -20,7 +21,7 @@ public class Offer {
 
     private String name;
 
-    private Integer cost;
+    private BigDecimal cost;
 
     private String description;
